+++ valhalla/common/rise_set_utils.py
@@ -1,10 +1,12 @@
 from math import cos, radians
 from datetime import timedelta
-from rise_set.astrometry import make_ra_dec_target, make_satellite_target, make_minor_planet_target, make_comet_target
+from rise_set.astrometry import make_ra_dec_target, make_satellite_target, make_minor_planet_target
+from rise_set.astrometry import make_comet_target, make_major_planet_target
 from rise_set.angle import Angle
 from rise_set.rates import ProperMotion
 from rise_set.utils import coalesce_adjacent_intervals
 from rise_set.visibility import Visibility
+from rise_set.moving_objects import MovingViolation
 from django.core.cache import cache
 
 from valhalla.common.configdb import configdb
@@ -44,14 +46,16 @@ def get_rise_set_intervals(request_dict, site=''):
             rise_set_target = get_rise_set_target(request_dict['target'])
             for window in request_dict['windows']:
                 visibility = get_rise_set_visibility(rise_set_site, window['start'], window['end'], site_details[site])
-
-                intervals_by_site[site].extend(
-                    visibility.get_observable_intervals(
-                        rise_set_target,
-                        airmass=request_dict['constraints']['max_airmass'],
-                        moon_distance=Angle(degrees=request_dict['constraints']['min_lunar_distance'])
+                try:
+                    intervals_by_site[site].extend(
+                        visibility.get_observable_intervals(
+                            rise_set_target,
+                            airmass=request_dict['constraints']['max_airmass'],
+                            moon_distance=Angle(degrees=request_dict['constraints']['min_lunar_distance'])
+                        )
                     )
-                )
+                except MovingViolation:
+                    pass
             intervals.extend(intervals_by_site[site])
     if request_dict.get('id'):
         cache.set(cache_key, intervals_by_site, 86400 * 30)  # cache for 30 days
@@ -88,7 +92,7 @@ def get_rise_set_target(target_dict):
                                             eccentricity=target_dict['eccentricity'],
                                             mean_anomaly=target_dict['meananom']
                                             )
-        else:
+        elif target_dict['scheme'] == 'MPC_COMET':
             return make_comet_target(target_type=target_dict['scheme'],
                                      epoch=target_dict['epochofel'],
                                      epochofperih=target_dict['epochofperih'],
@@ -98,6 +102,21 @@ def get_rise_set_target(target_dict):
                                      perihdist=target_dict['perihdist'],
                                      eccentricity=target_dict['eccentricity'],
                                      )
+        elif target_dict['scheme'] == 'JPL_MAJOR_PLANET':
+            return make_major_planet_target(target_type=target_dict['scheme'],
+                                            epochofel=target_dict['epochofel'],
+                                            inclination=target_dict['orbinc'],
+                                            long_node=target_dict['longascnode'],
+                                            arg_perihelion=target_dict['argofperih'],
+                                            semi_axis=target_dict['meandist'],
+                                            eccentricity=target_dict['eccentricity'],
+                                            mean_anomaly=target_dict['meananom'],
+                                            dailymot=target_dict['dailymot']
+                                            )
+        else:
+            raise TypeError('Invalid scheme ' + target_dict['scheme'])
+    else:
+        raise TypeError('Invalid target type' + target_dict['type'])
 
 
 def get_rise_set_site(site_detail):